+++ src/lib/FormattedInput.tsx
@@ -1,6 +1,6 @@
 import classnames from "classnames";
 import { Controller, FieldValues } from "react-hook-form";
-import { NumberFormatValues, NumericFormat, NumericFormatProps, PatternFormat, PatternFormatProps } from "react-number-format";
+import { NumericFormat, NumericFormatProps, PatternFormat, PatternFormatProps } from "react-number-format";
 import { useSafeNameId } from "src/lib/hooks/useSafeNameId";
 import { FormGroupLayout } from "./FormGroupLayout";
 import { CommonInputProps } from "./types/CommonInputProps";
@@ -11,7 +11,6 @@ interface FormattedInputProps<T extends FieldValues> extends CommonInputProps<T>
   patternFormat?: PatternFormatProps;
   numericFormat?: NumericFormatProps;
   placeholder?: string;
-  isAllowed?: ((values: NumberFormatValues) => boolean) | undefined
 }
 
 const FormattedInput = <T extends FieldValues>(props: FormattedInputProps<T>) => {
@@ -36,7 +35,6 @@ const FormattedInput = <T extends FieldValues>(props: FormattedInputProps<T>) =>
     className = "",
     hideValidationMessage = false,
     placeholder,
-    isAllowed,
   } = props;
   const { name, id } = useSafeNameId(props.name, props.id);
   const { control, disabled: formDisabled } = useFormContext();
@@ -94,7 +92,6 @@ const FormattedInput = <T extends FieldValues>(props: FormattedInputProps<T>) =>
                   onFocus={focusHandler}
                   style={style}
                   placeholder={placeholder}
-                  isAllowed={isAllowed}
                 ></NumericFormat>
               )}